Move upload_file Method
